Refactor: move catalog session management from UnityCatalogSamplerInterface to DatabricksSamplerInterface
