+++ ingestion/src/metadata/sampler/sqlalchemy/unitycatalog/sampler.py
@@ -14,8 +14,6 @@ Interfaces with database for all database engine
 supporting sqlalchemy abstraction layer
 """
 
-from sqlalchemy import event
-from sqlalchemy.orm import scoped_session, sessionmaker
 
 from metadata.sampler.sqlalchemy.databricks.sampler import DatabricksSamplerInterface
 
@@ -27,15 +25,3 @@ class UnityCatalogSamplerInterface(DatabricksSamplerInterface):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
-        # Create custom session with after_begin event to set catalog
-        session_maker = sessionmaker(bind=self.connection)
-
-        @event.listens_for(session_maker, "after_begin")
-        def set_catalog(session, transaction, connection):
-            connection.execute(
-                "USE CATALOG %(catalog)s;",
-                {"catalog": self.service_connection_config.catalog},
-            )
-
-        self.session_factory = scoped_session(session_maker)
